fix workout field types

+++ boot/src/main/kotlin/org/freekode/tp2intervals/infrastructure/trainingpeaks/workout/TPWorkoutDTO.kt
@@ -10,7 +10,7 @@ class TPWorkoutDTO(
     val workoutTypeValueId: Int?,
     val title: String,
     val totalTimePlanned: Double?,
-    val tssPlanned: Double?,
+    val tssPlanned: Int?,
     val description: String?,
     val coachComments: String?,
     val structure: TPWorkoutStructureDTO?
